@@ -20,12 +20,12 @@
     "dayjs": "^1.11.5",
     "destyle.css": "^3.0.2",
     "i18next": "^23.11.2",
-    "next": "^14.0.4",
+    "next": "^14.2.10",
     "next-i18next": "^15.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.48.2",
-    "react-i18next": "^14.1.0",
+    "react-i18next": "^14.1.2",
     "react-intersection-observer": "^9.4.4",
     "react-slick": "^0.29.0",
     "sharp": "^0.31.0",
@@ -37,13 +37,13 @@
     "@types/react": "^18.2.48",
     "@types/react-dom": "^18.2.18",
     "autoprefixer": "^10.4.12",
-    "eslint": "8.24.0",
-    "eslint-config-next": "12.3.1",
+    "eslint": "8.57.0",
+    "eslint-config-next": "14.2.10",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-unused-imports": "^2.0.0",
     "nodemailer": "^6.9.7",
     "postcss": "^8.4.18",
     "tailwindcss": "^3.2.1",
-    "typescript": "4.8.3"
+    "typescript": "5.6.2"
   }
 }
