@@ -1,20 +1,10 @@
-import type { NextPage } from "next";
+"use client";
 import Link from "next/link";
-import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 
-import Layout from "../components/layouts/Layout";
 import MainVisual from "../components/objects/organisms/MainVisual";
 import URL from "../const/url";
 
-export async function getStaticProps({ locale }: { locale: string }) {
-  return {
-    props: {
-      ...(await serverSideTranslations(locale, ["common"])),
-    },
-  };
-}
-
-const Home: NextPage = () => {
+export default function HomePage() {
   const tile = (title: string, image: string, link: string) => {
     return (
       <Link href={link}>
@@ -33,15 +23,14 @@ const Home: NextPage = () => {
   };
 
   return (
-    <Layout title="ARES Project" withSponsor>
+    <>
       <MainVisual />
       <section className="py-16 px-2 flex flex-col items-center gap-10 text-lg">
         <div className="border w-3/4 md:w-2/3 p-4 md:p-8 rounded-md">
           <h3 className="text-xl md:text-2xl">募金ご協力のお願い</h3>
           <p className="pt-2 pb-4 text-sm md:text-base">
             ARES
-            Projectでは、現在活動資金のためのご寄附を募集しております。皆様から頂いた資金は、機体の開発や科学分析に必要な物品の購入および大会や展示会の渡航費等に活用いたします。
-            皆様のご支援を心よりお待ちしております。
+            Projectでは、現在活動資金のためのご寄附を募集しております。皆様から頂いた資金は、機体の開発や科学分析に必要な物品の購入および大会や展示会の渡航費等に活用いたします。皆様のご支援を心よりお待ちしております。
           </p>
           <a href={URL.tohoku_university_fund} target="_blank" rel="noreferrer">
             <button className="bg-blue-500 text-white px-4 py-2 rounded-md text-sm md:text-base">
@@ -90,8 +79,6 @@ const Home: NextPage = () => {
           </div>
         </div>
       </section>
-    </Layout>
+    </>
   );
-};
-
-export default Home;
+}
