@@ -1,9 +1,10 @@
 /** @type {import('next').NextConfig} */
-const { i18n } = require('./next-i18next.config')
 const nextConfig = {
-  i18n,
   reactStrictMode: true,
   swcMinify: true,
+  compiler: {
+    emotion: true,
+  },
 }
 
 module.exports = nextConfig
